--- CD/lob1.cpp
@@ -1,1 +1,0 @@
-
